requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
databricks-sdk>=0.97.0
